requirements.txt
openai
pydantic
python-dotenv
anthropic
flask